+++ castle/cms/behaviors/contentredirectable.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+from castle.cms.utils import add_portal_message
 from plone.autoform import directives
 from plone.autoform.interfaces import IFormFieldProvider
 from plone.supermodel import model
@@ -48,11 +49,12 @@ class IContentRedirectable(model.Schema):
             u'If "Enable Redirect" is checked above, a redirect will be performed when a '
             u'non-editor attempts to navigate to this content. '
             u'The link is used almost verbatim, relative links become absolute and the strings '
-            u'"${navigation_root_url}" and "${portal_url}" get replaced with the real navigation_root_url or portal_url. '
-            u'If in doubt which one to use, please use navigation_root_url.'
+            u'"${navigation_root_url}" and "${portal_url}" get replaced with the real navigation_root_url '
+            u'or portal_url. If in doubt which one to use, please use navigation_root_url.'
         ),
         required=False,
     )
+
 
 @implementer(IContentRedirectable)
 @adapter(IDexterityContent)
@@ -134,7 +136,6 @@ class ContentRedirectable(object):
          - is_redirect_configured property returns True
          - the link is of a redirectable type (no mailto:, etc)
          - AND current user doesn't have permission to edit the Content Object"""
-
         if self.is_redirectable:
             url = self.absolute_target_url()
             if not url:
@@ -144,10 +145,10 @@ class ContentRedirectable(object):
                     'This object is currently configured to redirect to {}. '.format(url) +
                     'You are able to see this view because you have permission to edit this object.'
                 )
-                api.portal.show_message(
-                    message,
+                add_portal_message(
+                    message=message,
+                    message_type='warning',
                     request=self.request,
-                    type='warning',
                 )
             else:
                 return self.request.response.redirect(url.encode('utf-8'))